Reorganized Verification Process APIs and Updated Cancel Verification Button to Use Danger Styling
