+++ apps/meteor/app/livechat/server/api/rest.ts
@@ -12,3 +12,4 @@ import './v1/transfer';
 import './v1/contact';
 import './v1/webhooks';
 import './v1/integration';
+import './v1/verificationProcess';
