+++ apps/meteor/app/livechat/server/api/v1/visitor.ts
@@ -1,4 +1,3 @@
-import { OmnichannelVerification } from '@rocket.chat/core-services';
 import type { IRoom } from '@rocket.chat/core-typings';
 import { LivechatVisitors as VisitorsRaw, LivechatCustomField, LivechatRooms } from '@rocket.chat/models';
 import { Match, check } from 'meteor/check';
@@ -199,16 +198,3 @@ API.v1.addRoute('livechat/visitor.status', {
 		return API.v1.success({ token, status });
 	},
 });
-
-API.v1.addRoute(
-	'livechat/visitor.verify',
-	{ authRequired: true, permissionsRequired: ['initiate-livechat-verification-process'] },
-	{
-		async post() {
-			const { rid } = this.bodyParams;
-			await OmnichannelVerification.initiateVerificationProcess(rid);
-
-			return API.v1.success({ rid });
-		},
-	},
-);
